# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Millitary service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/commander/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new commander",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Commander"
                ],
                "summary": "Create Commander",
                "parameters": [
                    {
                        "description": "Commander Request",
                        "name": "CommanderReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/soldiers.CreateCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while creating",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/commander/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an existing commander",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Commander"
                ],
                "summary": "Delete Commander",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Commander ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while deleting",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/commander/get/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get an existing commander by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Commander"
                ],
                "summary": "Get Commander",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Commander ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Successful",
                        "schema": {
                            "$ref": "#/definitions/soldiers.Commander"
                        }
                    },
                    "401": {
                        "description": "Error while getting",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/commander/getall": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all commanders",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Commander"
                ],
                "summary": "Get All Commanders",
                "parameters": [
                    {
                        "type": "string",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get All Successful",
                        "schema": {
                            "$ref": "#/definitions/soldiers.AllCommanders"
                        }
                    },
                    "401": {
                        "description": "Error while getting all",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/commander/update/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing commander",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Commander"
                ],
                "summary": "Update Commander",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Commander ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Commander",
                        "name": "Commander",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/soldiers.Commander"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while updating",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/department/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new department",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "Create Department",
                "parameters": [
                    {
                        "description": "Department",
                        "name": "Department",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/soldiers.CreateDeportment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while creating",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/department/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an existing department",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "Delete Department",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Department ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while deleting",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/department/get/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get an existing department by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "Get Department",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Department ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Successful",
                        "schema": {
                            "$ref": "#/definitions/soldiers.Department"
                        }
                    },
                    "401": {
                        "description": "Error while getting",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/department/getall": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all departments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "Get All Departments",
                "parameters": [
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get All Successful",
                        "schema": {
                            "$ref": "#/definitions/soldiers.AllDepartments"
                        }
                    },
                    "401": {
                        "description": "Error while getting all",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/department/update/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing department",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "Update Department",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Department ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Department",
                        "name": "Department",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/soldiers.Department"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while updating",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "soldiers.AllCommanders": {
            "type": "object",
            "properties": {
                "commanders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/soldiers.Commander"
                    }
                }
            }
        },
        "soldiers.AllDepartments": {
            "type": "object",
            "properties": {
                "departments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/soldiers.Department"
                    }
                }
            }
        },
...
        "soldiers.Commander": {
            "type": "object",
            "properties": {
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                }
            }
        },
        "soldiers.CreateCommand": {
            "type": "object",
            "properties": {
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "soldiers.CreateDeportment": {
            "type": "object",
            "properties": {
                "commanders_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "soldiers.Department": {
            "type": "object",
            "properties": {
                "commander": {
                    "$ref": "#/definitions/soldiers.Commander"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authourization",
            "in": "header"
        }
    }
}